+++ samples/Cdr.Banking/Cdr.Banking.Business/Validation/TransactionArgsValidator.cs
@@ -16,7 +16,7 @@ namespace Cdr.Banking.Business.Validation
         {
             // Default FromDate where not provided, as 90 days less than ToDate; where no ToDate then assume today (now). Make sure FromDate is not greater than ToDate.
             Property(x => x.FromDate)
-                .Default(a => (a.ToDate!.HasValue ? a.ToDate.Value : DateTime.Now).AddDays(-90))
+                .Default(a => (a.ToDate!.HasValue ? a.ToDate.Value : DateTime.UtcNow).AddDays(-90))
                 .CompareProperty(CompareOperator.LessThanEqual, y => y.ToDate).DependsOn(y => y.ToDate);
 
             // Make sure MinAmount is not greater than MaxAmount.