+++ src/Beef.Data.Cosmos/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 Represents the **NuGet** versions.
 
+## v4.1.3
+- *Removed:* The `ICosmosDbArgs.SetIdentifierOnCreate` has been deprecated (removed). The new `IIdentifierGenerator` capability should be leveraged as a more flexible alternative.
+
 ## v4.1.2
 - *Fixed:* A `NotFoundException` will be thrown on a delete if it does not exist; otherwise, the application will assume it deleted successfully and the likes of a related event could be raised incorrectly.